Bump version to 2.2.6 (#64)
* distributed instance requests (#59)

* intellij suggestions

* remove jquery-ui dir
list.addAll replacement

* pom library updates

* Fix missing stylesheet in docs

* Library Updates

* ProjectDao cleanup

* Revert "ProjectDao cleanup"

This reverts commit cb843528c57e943f66092f3b60fc960c6f55992b.

* ProjectDao cleanup (2)

* Less config file reads

* removed some deprecated closeQuitely

* Never return an empty number of users
SRE-3170

* jobRegion.users required=true
SRE-3170

* validate number between 0-999999 users
SRE-3170

* require integer for initial users and increment
SRE-3170

* larger number of users

* Logging Instance Requests

* Testing distributed instance requests

* Testing distributed instance requests (1)

* Testing distributed instance requests (2)

* Reduce logging

* Randomly choose first subnet

* Request Agent AMI from AWS SSM

* Request Agent AMI from AWS SSM (1)

* Request Agent AMI from AWS SSM (2)

* Request Agent AMI from AWS SSM (3)

* Request Agent AMI from AWS SSM (4)

* PrimeFaces 6.2 (#60)

* PrimeFaces 6.2 update

* upgrade weld

* Up Versions

* wrap f:viewAction in f:metadata

* LogViewer

* Debug Log Viewer

* Fix CodeMirror javascript

* works this way

* Remove ContextType definition
because: SEVERE [http-nio-8080-exec-6] com.sun.faces.context.AjaxExceptionHandlerImpl.handlePartialResponseError java.lang.IllegalArgumentException: Unrecognized Content Type.

* Resolve RES_NOT_FOUND for css

* Resolve RES_NOT_FOUND for css (2)

* Resolve RES_NOT_FOUND for css (3)

* Resolve RES_NOT_FOUND for css (4)

* Resolve RES_NOT_FOUND for css (5)

* Resolve RES_NOT_FOUND for css (6)

* Resolve RES_NOT_FOUND for css (7)

* Reset back to Production

* Update graphicImage to use library tag

* consistent JSF library use

* consistent JSF library use (2)

* use more jquery icons

* use more jquery icons (2)

* use more jquery icons (3)

* use more jquery icons (4)

* use more jquery icons (5)

* update faces-config

* add redirect tag

* use more jQuery icons (6)

* update buildspec to 0.2

* Empty commit to trigger post_receive hooks

* Library updates (#61)

* update pom dependencies

* jUnit 5 changes

* Fix proxy recorder by rolling back commons-collections version
commons-configuration has a dependency on the old version

* Cleanup up PickList usage

* Revert to Primefaces 6.1

* Remove wagon-webdav-jackrabbit
AWS codeBuild will not resolve the dependency , ReasonPhrase:Forbidden.

* Restore Filtering

* Redirect for reapply-filters.xhtml

* adding logo to be used by the OS Site (#62)

* Json logging (#63)

* jsonlayout logging

* Consistent JSON logging
beyond the ThreadLocalLogEvent

* Verbose logging on Validation failures

* more logUtil usage

* Update logging for consistency

* Add guava lib to the war deployment

* Update lucene 3.6.2 -> 4.10.4

* more json logging

* one more logging change

* close httpclient one thread completion, or kill

* Bump version to 2.2.6

+++ web/web_support/src/main/java/com/intuit/tank/project/JobValidator.java
@@ -121,17 +121,17 @@ public class JobValidator {
 
     public long getDurationMs(String group) {
         Long duration = expectedDuration.get(group);
-        return duration != null ? duration.longValue() : 0L;
+        return duration != null ? duration : 0L;
     }
 
     public String getDuration(String group) {
         Long duration = expectedDuration.get(group);
-        return TimeUtil.toTimeString(duration != null ? duration.longValue() : 0);
+        return TimeUtil.toTimeString(duration != null ? duration : 0);
     }
 
     public long getExpectedTime(String group) {
         Long duration = expectedDuration.get(group);
-        return duration != null ? duration.longValue() : 0;
+        return duration != null ? duration : 0;
     }
 
     private void putVariable(Map<String, Set<String>> map, String key, String value, String declarer) {
@@ -233,7 +233,7 @@ public class JobValidator {
 
     private void addDuration(String groupName, int loop, long milis) {
         Long val = this.expectedDuration.get(groupName);
-        long l = val != null ? val.longValue() : 0L;
+        long l = val != null ? val : 0L;
         l += milis * loop;
         expectedDuration.put(groupName, l);
     }
